Challenge #2 - One small step at a time
Instructions

Click here to learn how to navigate the code editor

Registering new astronauts was handled manually in previous missions but it's taking too much time. Knowing that the astronaut information is already listed in a data set, you propose to optimize that job with programming.

Your task is to automate that process by creating a function that will take in an astronaut object as a parameter and return a string using the properties of the astronaut.