Challenge #14 - List-o-mania
(This challenge is worth 10 points)

Your task is to write a function that will take in an array of objects containing switches. The function will change the value of the isOn property true for every switch in the list.

You can use the previous switchToggle function from the fourth challenge, however be careful, since we want all of them on, not toggled!